requirements.txt
aws-cdk-lib==2.148.0
constructs>=10.0.0,<11.0.0
pytest==8.2.2
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.8
